fixed circular inclusions

+++ Parameters/DefaultValues.h
@@ -7,21 +7,21 @@ static const struct PitchShifterDefaults
 {
 	float bypass = 1;
 	float amount = 0.5;
-	int semitones = 12;
+	float semitones = 12;
 	float rand = 0;
 } pitchShifterDefs;
 
 static const struct WaveshaperDefaults
 {
-	int bypass = 1;
+	float bypass = 1;
     float input = 0.5;
-    int waveshape = 0;
+    float waveshape = 0;
     float amount = 0.5;
 } waveshaperDefs;
 
 static const struct ReverbDefaults
 {
-	int bypass = 1;
+	float bypass = 1;
 	float amount = 0.5;
 	float mode = 0;
 	float size = 0.5;
